Actividad 8/requirements.txt
flask
redis
python-dotenv
celery
flask-mail